requirements.txt
dnspython==2.3.0
pyfiglet==0.8.post1
colorama==0.4.6
tabulate==0.9.0
